fix: 优化Worker信息显示,更新为不可选中建筑,解决Worker Seek时接取任务出现直接不移动直接Carry的问题
