+++ Scripts/2D/UI/Panel/BackpackMenuPanel.cs
@@ -13,7 +13,7 @@
         {
             this.Name = "BackpackMenu";
             this.Select = new SelectItemData();
-            this.OpenPanel();
+            this.Open();
             Tool.GetComponentInChildren<Button>(this.Panel, "Equip").onClick.AddListener(this.OnClick_Equip);
             Tool.GetComponentInChildren<Button>(this.Panel, "Abandon").onClick.AddListener(this.OnClick_Abandon);
             Tool.GetComponentInChildren<Button>(this.Panel, "BackGame").onClick.AddListener(this.OnClick_BackGame);
