+++ Scripts/2D/UI/Panel/CreateDataPanel.cs
@@ -17,7 +17,7 @@
         public CreateDataPanel()
         {
             this.Name = "CreateData";
-            this.OpenPanel();
+            this.Open();
             Transform g1 = Tool.GetComponentInChildren<Transform>(this.Panel, "MapHeight");
             Slider s1 = g1.Find("Bar").GetComponent<Slider>();
             this.height = (int)s1.value;
