+++ Scripts/2D/UI/Panel/CreateMenuPanel.cs
@@ -12,7 +12,7 @@
         public CreateMenuPanel()
         {
             this.Name = "CreateMenu";
-            this.OpenPanel();
+            this.Open();
             Tool.GetComponentInChildren<Button>(this.Panel, "StartCreate").onClick.AddListener(this.OnClick_StartCreate);
             Tool.GetComponentInChildren<Button>(this.Panel, "Back").onClick.AddListener(this.OnClick_Back);
         }
