+++ Scripts/2D/UI/Panel/JoinMenuPanel.cs
@@ -13,7 +13,7 @@
         public JoinMenuPanel()
         {
             this.Name = "JoinMenu";
-            this.OpenPanel();
+            this.Open();
             Tool.GetComponentInChildren<Button>(this.Panel, "StartJoin").onClick.AddListener(this.OnClick_StartJoin);
             Tool.GetComponentInChildren<Button>(this.Panel, "Back").onClick.AddListener(this.OnClick_Back);
         }
